Feat: 편지 봉투 파일 병합, 편지함 분리
